+++ Game/Game/PatrolState.h
@@ -11,14 +11,14 @@ public:
 
 	PatrolState() : currentPoint(0)
 	{}
-	PatrolState(std::vector<Vector2i> points, DynamicObject *_owner, int currPoint)
+	PatrolState(std::vector<Vector2f> points, DynamicObject *_owner, int currPoint)
 	{
 		patrolPoints	= points;
 		owner			= _owner;
 		currentPoint	= currPoint;
 		pathOffset		= 0;
 	}
-	PatrolState(std::vector<Vector2i> points, DynamicObject *_owner, int currPoint, int _pathOffset)
+	PatrolState(std::vector<Vector2f> points, DynamicObject *_owner, int currPoint, int _pathOffset)
 	{
 		patrolPoints	= points;
 		owner			= _owner;
@@ -38,7 +38,7 @@ public:
 		//Patrolling
 		if (currentPoint < patrolPoints.size() - 1)
 		{
-			Vector2i point = patrolPoints[currentPoint] - owner->GetPos();				//Находим направление в котором надо идти
+			Vector2f point = patrolPoints[currentPoint] - owner->GetPos();				//Находим направление в котором надо идти
 			double vecLength = sqrt(point.x * point.x + point.y * point.y);
 			if (vecLength <= pathOffset)												//Если мы уже достигли точки
 			{
@@ -47,8 +47,8 @@ public:
 			}
 			else
 			{
-				Vector2i normalizedDir = Vector2i(point.x / vecLength, point.y / vecLength);		//Нормализованное направление
-				owner->SetPos(Vector2i(owner->GetPos().x + normalizedDir.x,							//Устанавливаем новую позицию нпс
+				Vector2f normalizedDir = Vector2f(point.x / vecLength, point.y / vecLength);		//Нормализованное направление
+				owner->SetPos(Vector2f(owner->GetPos().x + normalizedDir.x,							//Устанавливаем новую позицию нпс
 										owner->GetPos().y + normalizedDir.y));
 			}
 		}
@@ -64,7 +64,7 @@ public:
 	}
 
 private:
-	std::vector<Vector2i>	patrolPoints;
+	std::vector<Vector2f>	patrolPoints;
 	DynamicObject			*owner;
 	int						pathOffset;
 	int						currentPoint;
